+++ packages/flame/CHANGELOG.md
@@ -8,6 +8,8 @@
  - Moving tutorials to the Flame main repository
  - Fixing Flutter and Dart version constraints
  - Exporting Images and AssetsCache
+ - Make Size and Position in PositionComponent final
+ - Add snapTo and onPositionUpdate to Camera
 
 ## 1.0.0-rc8
  - Migrate to null safety